+++ ts/features/itwallet/presentation/remote/machine/__tests__/machine.test.ts
@@ -72,8 +72,8 @@ describe("itwRemoteMachine", () => {
     actor.send({
       type: "start",
       payload: {
-        clientId: T_CLIENT_ID,
-        requestUri: T_REQUEST_URI,
+        client_id: T_CLIENT_ID,
+        request_uri: T_REQUEST_URI,
         state: T_STATE
       } as ItwRemoteRequestPayload
     });
@@ -90,8 +90,8 @@ describe("itwRemoteMachine", () => {
       value: "Failure",
       context: {
         payload: {
-          clientId: T_CLIENT_ID,
-          requestUri: T_REQUEST_URI,
+          client_id: T_CLIENT_ID,
+          request_uri: T_REQUEST_URI,
           state: T_STATE
         } as ItwRemoteRequestPayload
       }
@@ -114,8 +114,8 @@ describe("itwRemoteMachine", () => {
       value: "Failure",
       context: {
         payload: {
-          clientId: T_CLIENT_ID,
-          requestUri: T_REQUEST_URI,
+          client_id: T_CLIENT_ID,
+          request_uri: T_REQUEST_URI,
           state: T_STATE
         } as ItwRemoteRequestPayload
       }
@@ -138,8 +138,8 @@ describe("itwRemoteMachine", () => {
       value: "Failure",
       context: {
         payload: {
-          clientId: T_CLIENT_ID,
-          requestUri: T_REQUEST_URI,
+          client_id: T_CLIENT_ID,
+          request_uri: T_REQUEST_URI,
           state: T_STATE
         } as ItwRemoteRequestPayload
       }
@@ -164,8 +164,8 @@ describe("itwRemoteMachine", () => {
     actor.send({
       type: "start",
       payload: {
-        clientId: T_CLIENT_ID,
-        requestUri: T_REQUEST_URI,
+        client_id: T_CLIENT_ID,
+        request_uri: T_REQUEST_URI,
         state: T_STATE
       } as ItwRemoteRequestPayload
     });
@@ -184,8 +184,8 @@ describe("itwRemoteMachine", () => {
     actor.send({
       type: "start",
       payload: {
-        clientId: T_CLIENT_ID,
-        requestUri: T_REQUEST_URI,
+        client_id: T_CLIENT_ID,
+        request_uri: T_REQUEST_URI,
         state: T_STATE
       } as ItwRemoteRequestPayload
     });